agrisakhi_backend/requirements.txt
cloudinary==1.44.2
Django==4.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-filter==25.1
django-storages==1.14.6
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
gunicorn==26.0.0
pillow==12.2.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
